+++ app/Http/Kernel.php
@@ -7,6 +7,7 @@ use App\Http\Middleware\LogContext;
 use App\Http\Middleware\RoomAuthenticate;
 use App\Http\Middleware\RouteEnableIfConfig;
 use App\Http\Middleware\RouteEnableIfSetting;
+use App\Http\Middleware\StoreSessionData;
 use Illuminate\Foundation\Http\Kernel as HttpKernel;
 use Laravel\Sanctum\Http\Middleware\EnsureFrontendRequestsAreStateful;
 
@@ -20,6 +21,7 @@ class Kernel extends HttpKernel
      * @var array
      */
     protected $middleware = [
+        \App\Http\Middleware\TrustHosts::class,
         \App\Http\Middleware\TrustProxies::class,
         \Illuminate\Http\Middleware\HandleCors::class,
         \App\Http\Middleware\PreventRequestsDuringMaintenance::class,
@@ -42,13 +44,16 @@ class Kernel extends HttpKernel
             \App\Http\Middleware\VerifyCsrfToken::class,
             \Illuminate\Routing\Middleware\SubstituteBindings::class,
             \App\Http\Middleware\SetApplicationLocale::class,
+            'shibboleth',
             'loggedin:ldap,users',
+            StoreSessionData::class,
             LogContext::class,
         ],
 
         'api' => [
             EnsureFrontendRequestsAreStateful::class,
             \Illuminate\Routing\Middleware\ThrottleRequests::class.':api',
+            'shibboleth',
             'loggedin:ldap,users',
             \Illuminate\Routing\Middleware\SubstituteBindings::class,
             \App\Http\Middleware\SetApplicationLocale::class,
@@ -78,5 +83,6 @@ class Kernel extends HttpKernel
         'check.stale'           => EnsureModelNotStale::class,
         'enable_if_config'      => RouteEnableIfConfig::class,
         'enable_if_setting'     => RouteEnableIfSetting::class,
+        'shibboleth'            => \App\Auth\Shibboleth\ShibbolethSessionMiddleware::class,
     ];
 }