+++ resources/js/views/VerifyEmail.vue
@@ -49,6 +49,7 @@ export default {
   mounted () {
     this.verifyEmail();
   },
+
   methods: {
     verifyEmail () {
       this.loading = true;