+++ CHANGELOG.md
@@ -6,10 +6,15 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 ### Added
-- Support for postgres database on db:import command ([#545])
+- Shibboleth authentication ([#20], [#497])
+- Database Import command: Support for postgres database ([#545])
+- Greenlight v2 migration command: Import social users ([#497])
+- Greenlight v2 migration command: Using database transactions to undo changes in case of errors ([#497])
 
 ### Fixed
 - Missing locale strings from fallback locale ([#576])
+- Greenlight v2 migration command: Room feature 'Room Authentication' missing ([#497])
+- Greenlight v2 migration command: Room name too long ([#497])
 - Room type unselected in room settings on changes to the room type settings ([#577], [#586])
 
 ### Changed
@@ -413,6 +418,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 [#17]: https://github.com/THM-Health/PILOS/pull/17
 [#18]: https://github.com/THM-Health/PILOS/pull/18
 [#19]: https://github.com/THM-Health/PILOS/issues/19
+[#20]: https://github.com/THM-Health/PILOS/issues/20
 [#21]: https://github.com/THM-Health/PILOS/pull/21
 [#22]: https://github.com/THM-Health/PILOS/pull/22
 [#30]: https://github.com/THM-Health/PILOS/issues/30
@@ -638,6 +644,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 [#389]: https://github.com/THM-Health/PILOS/pull/389
 [#469]: https://github.com/THM-Health/PILOS/pull/469
 [#470]: https://github.com/THM-Health/PILOS/pull/470
+[#497]: https://github.com/THM-Health/PILOS/pull/497
 [#503]: https://github.com/THM-Health/PILOS/issues/503
 [#504]: https://github.com/THM-Health/PILOS/pull/504
 [#505]: https://github.com/THM-Health/PILOS/issues/505